requirements.txt
# Django and Core Dependencies
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
# psycopg2-binary==2.9.7  # Commented out for now, using SQLite
dj-database-url==2.1.0

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0

# Media & File Handling
Pillow==10.0.1
django-storages==1.14.2
boto3==1.28.85

# Email & Notifications
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Admin & CMS
django-admin-interface==0.26.1
django-ckeditor==6.7.0
django-mptt==0.15.0

# API Documentation
drf-spectacular==0.26.5

# Development Tools
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
requests==2.31.0
